Issue #145 chore: rename interface
