+++ app/src/main/java/com/gdsc/sogongsogong/data/remote/post/PostDataSource.kt
@@ -1,9 +1,9 @@
-package com.gdsc.sogongsogong.data.remote
+package com.gdsc.sogongsogong.data.remote.post
 
 import com.gdsc.sogongsogong.data.entity.Post
 import kotlinx.coroutines.flow.Flow
 
-interface PostRemoteDataSource {
+interface PostDataSource {
 
     suspend fun fetchPost(postId: Long): Flow<Post>
