+++ app/src/main/java/com/gdsc/sogongsogong/di/module/DataSourceModule.kt
@@ -1,7 +1,7 @@
 package com.gdsc.sogongsogong.di.module
 
-import com.gdsc.sogongsogong.data.remote.PostRemoteDataSource
-import com.gdsc.sogongsogong.data.remote.PostRemoteDataSourceImpl
+import com.gdsc.sogongsogong.data.remote.post.PostDataSource
+import com.gdsc.sogongsogong.data.remote.post.PostRemoteDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -12,5 +12,5 @@ import dagger.hilt.components.SingletonComponent
 abstract class DataSourceModule {
 
     @Binds
-    abstract fun bindPostDataSource(dataSource: PostRemoteDataSourceImpl): PostRemoteDataSource
+    abstract fun bindPostDataSource(dataSource: PostRemoteDataSourceImpl): PostDataSource
 }
