+++ app/src/main/java/com/gdsc/sogongsogong/ui/board/BoardViewModel.kt
@@ -3,7 +3,7 @@ package com.gdsc.sogongsogong.ui.board
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import com.gdsc.sogongsogong.di.dispatcher.DispatcherProvider
-import com.gdsc.sogongsogong.data.remote.PostRemoteDataSource
+import com.gdsc.sogongsogong.data.remote.post.PostDataSource
 import com.gdsc.sogongsogong.ui.base.BaseViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
@@ -13,7 +13,7 @@ import javax.inject.Inject
 @HiltViewModel
 class BoardViewModel @Inject constructor(
     dispatcherProvider: DispatcherProvider,
-    private val postRemoteDataSource: PostRemoteDataSource
+    private val postDataSource: PostDataSource
 ) : BaseViewModel(dispatcherProvider) {
 
     private var _board: MutableLiveData<String> = MutableLiveData() // FIXME: entity로 변경
@@ -24,7 +24,7 @@ class BoardViewModel @Inject constructor(
 
     fun fetchAllPost(page: Int) {
         onIo {
-            postRemoteDataSource.fetchAllPost(page.toLong())
+            postDataSource.fetchAllPost(page.toLong())
         }
     }
